Change the FurnitureID datatype to int64

+++ internal/data/furniture.go
@@ -3,7 +3,7 @@ package data
 // Furniture is a struct that holds information about
 // a specific furniture.
 type Furniture struct {
-	FurnitureID int
+	FurnitureID int64
 	Name        string
 	Description string
 	Price       float64
